+++ app/src/domain/adc_domain/hardware/adc_emulator.h
@@ -7,7 +7,7 @@ namespace eerie_leap::domain::adc_domain::hardware {
 class AdcEmulator : public Adc {
 public:
     int Initialize() override;
-    double ReadChannel(int channel) override;
+    float ReadChannel(int channel) override;
 };
 
 }  // namespace eerie_leap::domain::adc_domain::hardware
