+++ app/src/domain/adc_domain/hardware/i_adc.h
@@ -11,7 +11,7 @@ class IAdc {
 public:
     virtual int Initialize() = 0;
     virtual void UpdateConfiguration(AdcConfiguration config) = 0;
-    virtual double ReadChannel(int channel) = 0;
+    virtual float ReadChannel(int channel) = 0;
 };
 
 }  // namespace eerie_leap::domain::adc_domain::hardware
