+++ app/src/domain/sensor_domain/utilities/voltage_interpolator/i_voltage_interpolator.h
@@ -12,7 +12,7 @@ using namespace eerie_leap::domain::sensor_domain::models;
 
 class IVoltageInterpolator {
 public:
-    virtual double Interpolate(double voltage) const = 0;
+    virtual float Interpolate(float voltage) const = 0;
     virtual const std::shared_ptr<std::vector<CalibrationData>> GetCalibrationTable() const = 0;
     virtual const InterpolationMethod GetInterpolationMethod() const = 0;
 };
